requirements.txt
Django==1.9
pillow==3.1.0
django-pagedown==0.1.1
django-markdown-deux==1.0.5
markdown2==2.3.1
django-crispy-forms==1.6.0
djangorestframework==3.3.3
django-filter==0.13.0
djangorestframework-jwt==1.8.0